+++ src/sdk/CHANGELOG.md
@@ -51,3 +51,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Moved to complete async internal implementation [jansenbe - Bert Jansen]
 - Code cleanup (Teams implementation / ODataQuery) [JarbasHorst - Jarbas Horst]
 - When a non loaded model properties is assigned it now is considered as a change and send back to the server on update #106 [jansenbe - Bert Jansen]
+- Throw client exception when an API call still contains unresolved tokens before being executed #107 [jansenbe - Bert Jansen]